Script to forcefully match module BuildIDs
To be used with extreme caution. The purpose of this script is to make sure that the BuildID for the modules matches the BuildID for the engine.
This is to use plugins from the Epic Games Launcher made for 5.1 with a (pre)compiled engine based off of 5.1.
Ideally you'd want to set the BuildID of the (pre)compiled engine to match the BuildID of the Epic Games Launcher version of UE 5.1.